requirements.txt
psycopg2-binary==2.9.10
httpx==0.28.1
python-dotenv==1.0.1
